requirements.txt
fastapi
uvicorn
python-dotenv
pymongo
requests
pymupdf
fastapi-session
azure-cognitiveservices-speech
pydub